move research purpose query tests from automation to unit
